+++ lib/matplotlib/backends/_backend_gtk3.py
@@ -32,7 +32,7 @@ except ImportError:
 
 import matplotlib
 
-from ._backend_bases import (RendererBase, GraphicsContextBase,
+from .base_backend_bases import (RendererBase, GraphicsContextBase,
         FigureManagerBase, FigureCanvasBase,
         NavigationToolbar2, cursors, TimerBase)
