+++ lib/matplotlib/backends/base_backend_bases.py
